fix refactor f32 to i16 AB#507

+++ src/engine.rs
@@ -16,10 +16,10 @@ use wasm_bindgen::JsCast;
 use async_trait::async_trait;
 
 pub struct Rect {
-    pub x:f32,
-    pub y:f32,
-    pub width:f32,
-    pub height:f32,
+    pub x:i16,
+    pub y:i16,
+    pub width:i16,
+    pub height:i16,
 }
 impl Rect {
     pub fn intersects(&self, other: &Rect) -> bool {
@@ -369,8 +369,8 @@ impl Image {
         let bounding_box = Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: element.width() as f32,
-            height: element.height() as f32
+            width: element.width() as i16,
+            height: element.height() as i16
         };
         Self {
             element,
@@ -389,7 +389,7 @@ impl Image {
     }
 
     pub fn set_x(&mut self, x: i16) {
-        self.bounding_box.x = x as f32;
+        self.bounding_box.x = x ;
         self.position.x = x;
     }
     pub fn right(&self) -> i16 {
