+++ src/game.rs
@@ -123,10 +123,10 @@ impl Game for WalkTheDog {
     }
     fn draw(&self, renderer: &Renderer) {
         renderer.clear(&Rect {
-            x: 0.0,
-            y: 0.0,
-            width: HEIGHT as f32,
-            height: 600.0,
+            x: 0,
+            y: 0,
+            width: HEIGHT,
+            height: 600,
         });
         if let WalkTheDog::Loaded(walk) = self {
             walk.backgrounds.iter().for_each(|backgdound|{ backgdound.draw(renderer); });
@@ -163,18 +163,18 @@ impl RedHatBoy {
     fn destination_box(&self) -> Rect {
         let sprite = self.current_sprite().expect("Cell not found");
         Rect {
-            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x as i16)
+            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x)
                 .into(),
-            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y as i16)
+            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y)
                 .into(),
             width: sprite.frame.w.into(),
             height: sprite.frame.h.into(),
         }
     }
     fn bounding_box(&self) -> Rect {
-        const X_OFFSET: f32 = 18.0;
-        const Y_OFFSET: f32 = 14.0;
-        const WIDTH_OFFFSET: f32 = 28.0;
+        const X_OFFSET: i16 = 18;
+        const Y_OFFSET: i16 = 14;
+        const WIDTH_OFFFSET: i16 = 28;
         let mut bounding_box = self.destination_box();
         bounding_box.x += X_OFFSET;
         bounding_box.y += Y_OFFSET;
@@ -212,7 +212,7 @@ impl RedHatBoy {
     fn knock_out(&mut self) {
         self.state_machine = self.state_machine.transition(Event::KnockOut);
     }
-    fn land_on(&mut self, position: f32) {
+    fn land_on(&mut self, position: i16) {
         self.state_machine = self.state_machine.transition(Event::Land(position));
     }
     fn pos_y(&self) -> i16 {
@@ -381,9 +381,9 @@ mod red_hat_boy_states {
                 _state: Falling {},
             }
         }
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Running> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Running> {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Running {},
             }
         }
@@ -413,9 +413,9 @@ mod red_hat_boy_states {
                 _state: Falling {},
             }
         }
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Sliding> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Sliding> {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Sliding {},
             }
         }
@@ -433,9 +433,9 @@ mod red_hat_boy_states {
                 JumpingEndState::Jumping(self)
             }
         }
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Running> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Running> {
             RedHatBoyState {
-                context: self.context.reset_frame().set_on(position as i16),
+                context: self.context.reset_frame().set_on(position),
                 _state: Running {},
             }
         }
@@ -492,7 +492,7 @@ pub enum Event {
     Jump,
     KnockOut,
     Update,
-    Land(f32),
+    Land(i16),
 }
 #[derive(Copy, Clone)]
 enum RedHatBoyStateMachine {
@@ -656,8 +656,8 @@ impl Platform {
         }
     }
     fn bounding_boxes(&self) -> Vec<Rect> {
-        const X_OFFSET: f32 = 60.0;
-        const END_HEIGHT: f32 = 54.0;
+        const X_OFFSET: i16 = 60;
+        const END_HEIGHT: i16 = 54;
         let destination_box = self.destination_box();
         let bounding_box_one = Rect {
             x: destination_box.x,
@@ -668,7 +668,7 @@ impl Platform {
         let bounding_box_two = Rect {
             x: destination_box.x + X_OFFSET,
             y: destination_box.y,
-            width: destination_box.width - (X_OFFSET * 2.0),
+            width: destination_box.width - (X_OFFSET * 2),
             height: destination_box.height,
         };
         let bounding_box_three = Rect {
